requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.2
mysql_connector_repackaged==0.3.1
natasha==1.4.0
pymorphy2==0.9.1
requests==2.26.0
Telethon==1.23.0
tqdm==4.62.3
